fix: Messages of toastify are now showed
